+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [CalVer](https://calver.org/).
 - Feature info click wouldn't open the pop-up if multiple layers was switch on.
 - Drawings are not longer dashed, when selected.
 - Conflict search module will now search protected layers if user is signed in.
+- On `zoomToExtentOfgeoJsonStore` in geocloud.js pan map one pixel to defeat a strange bug, which causes a browser freeze.
 
 ## [2021.5.0] - 2021-4-5
 ### Changed